separate colored elements on title screen
