+++ art/title_screen.rb
@@ -43,15 +43,18 @@ def play_again
   ASCII
 end
 
-TITLE_SCREEN = # Title Screen
+CASTLE_CARNAGE =
   <<-'ASCII'
-
           ╭━━━┳━━━┳━━━┳━━━━┳╮╱╱╭━━━╮ ╭━━━┳━━━┳━━━┳━╮╱╭┳━━━┳━━━┳━━━╮
           ┃╭━╮┃╭━╮┃╭━╮┃╭╮╭╮┃┃╱╱┃╭━━╯ ┃╭━╮┃╭━╮┃╭━╮┃┃╰╮┃┃╭━╮┃╭━╮┃╭━━╯
           ┃┃╱╰┫┃╱┃┃╰━━╋╯┃┃╰┫┃╱╱┃╰━━╮ ┃┃╱╰┫┃╱┃┃╰━╯┃╭╮╰╯┃┃╱┃┃┃╱╰┫╰━━╮
           ┃┃╱╭┫╰━╯┣━━╮┃╱┃┃╱┃┃╱╭┫╭━━╯ ┃┃╱╭┫╰━╯┃╭╮╭┫┃╰╮┃┃╰━╯┃┃╭━┫╭━━╯
           ┃╰━╯┃╭━╮┃╰━╯┃╱┃┃╱┃╰━╯┃╰━━╮ ┃╰━╯┃╭━╮┃┃┃╰┫┃╱┃┃┃╭━╮┃╰┻━┃╰━━╮
           ╰━━━┻╯╱╰┻━━━╯╱╰╯╱╰━━━┻━━━╯ ╰━━━┻╯╱╰┻╯╰━┻╯╱╰━┻╯╱╰┻━━━┻━━━╯
+  ASCII
+
+CASTLE = # Title Screen
+  <<-'ASCII'
                                     |>>>
                                     |
                       |>>>      _  _|_  _         |>>>
@@ -71,7 +74,6 @@ TITLE_SCREEN = # Title Screen
         ____--`~    '--~~__|.    |+++++__|----~    ~`---,              ___
    -~--~                   ~---__|,--~'                  ~~----_____-~'   `~----~~
 
-                              ⲚⲀⲘⲈ ⲨⲞ𐌵ꞅϨⲈ𝓛𝓕
   ASCII
 
 # Main to Boss Transition
