+++ assets/ascii.rb
@@ -91,3 +91,4 @@ SHOP      = "#{GN}𝕋𝕙𝕖 𝕊𝕙𝕠𝕡#{CL}"
 GANKED    = "#{BL}ＧＡＮＫＥＤ#{CL}"
 WASTED    = "#{OR}𝕎𝔸𝕊𝕋𝔼𝔻#{CL}"
 LEVEL     = "#{BL}LEͤVͮEͤL UͧРⷬ#{CL}"
+NAME      = "#{YL}ⲚⲀⲘⲈ ⲨⲞ𐌵ꞅϨⲈ𝓛𝓕 𝓕ⲞⲞ𝓛#{CL}"
