+++ imports.rb
@@ -1,7 +1,7 @@
 # rubocop:disable all
 require_relative 'art/battlefield'
-require_relative 'art/games_display'
 require_relative 'art/room_service'
+require_relative 'art/screen'
 require_relative 'art/title_screen'
 require_relative 'assets/ascii'
 require_relative 'assets/cheat_mode'
